+++ src/app.module.ts
@@ -11,15 +11,12 @@ import { AuthzModule } from 'fintegrate-auth';
 import { TasksModule } from './tasks/module/tasks.module';
 import { TransformDataStructure } from './transformDataStructure/convertData';
 import { MessagesModule } from './message/module/messages.module';
-import { UserModule } from './user/module/users.module';
 
 @Module({
   imports: [
     ConfigModule.forRoot({
       isGlobal: true,
     }),
-
-    UserModule,
     TasksModule,
     WorkersModule,
     MessagesModule,
