--- src/user/services/users.service.spec.ts
@@ -1,54 +1,0 @@
-// import { Test, TestingModule } from '@nestjs/testing';
-// import { UserService } from './users.service';
-// import { getModelToken } from '@nestjs/mongoose';
-// import { Model } from 'mongoose';
-
-// describe('UsersService', () => {
-//   let service: UserService;
-//   let model: Model<any>
-
-//   beforeEach(async () => {
-//     const module: TestingModule = await Test.createTestingModule({
-//       providers: [UserService],
-//     }).compile();
-
-//     service = module.get<UserService>(UserService);
-//     model = module.get<Model<any>>(getModelToken('User'));
-//   });
-
-//   it('should be defined', () => {
-//     expect(service).toBeDefined();
-//   });
-
-//   it('should create a new user when createUser is called', async () => {
-//     const mockCreateUserDto = {
-//       userName: 'Test User',
-//       userEmail: 'test@example.com',
-//       auth0_user_id: 'n05y7452opu',
-//       registeredAt: new Date('2024-12-31'),
-//       lastLogin: new Date('2024-12-31'),
-//       mobile: '0555555555',
-//       status: 'Married',
-//       dateOfBirth: new Date('2024-12-31'),
-//       address: {
-//         city: 'TV',
-//         street: 'Hshalom',
-//         num: 5,
-//       },
-//     };
-//     const mockUserModel = {
-//       save: jest.fn(() => Promise.resolve(mockCreateUserDto)),
-//     };
-
-//     jest.spyOn(model, 'findOne').mockReturnValue(mockUserModel as any);
-//     const createdUser = await service.createUser(mockCreateUserDto);
-//     expect(createdUser).toEqual(mockCreateUserDto);
-//   });
-
-// });
-
-it('always returns true', () => {
-  expect(true).toBe(true);
-});
-
-
